homelab-weather/requirements.txt
pimoroni-bme280==1.0.0
python-telegram-bot==21.6
smbus2==0.5.0
prometheus_client==0.21.0
